Move nvtx3.hpp to include/nvtx3 directory.
