+++ cpp/tests/CMakeLists.txt
@@ -41,7 +41,7 @@ endfunction(ConfigureTest)
 
 include_directories("${GTEST_INCLUDE_DIR}"
                     "${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES}"
-                    "${CMAKE_SOURCE_DIR}")
+                    "${CMAKE_SOURCE_DIR}/include")
 
 ###################################################################################################
 # - library paths ---------------------------------------------------------------------------------
